requirements.txt
discord-py-slash-command >= 4.2.1
discord.py >= 2.4.0
psutil >= 5.9.8
colorama >= 0.4.6
